feat(util): add GetStatus & Merge functions
GetStatus
This function makes the network request for the given url.
It unmarshals the json response to a struct it gets from the
Response parameter. Response is an interface for pointers to
the different Response structs.
It returns an error channel.

Merge
This function fans in the different error channels into one.